+++ stroom-core-shared/src/main/java/stroom/security/shared/UserNameProvider.java
@@ -1,0 +1,8 @@
+package stroom.security.shared;
+
+import stroom.util.shared.ResultPage;
+
+public interface UserNameProvider {
+
+    ResultPage<String> findUserNames(FindUserNameCriteria criteria);
+}